# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AutoPro API",
    "description": "Documentacao das rotas da API AutoPro baseada na modelagem Cliente, Veiculo, Servico, Item de Servico e Produto.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor local"
    }
  ],
...
  "paths": {
...
    "/clientes": {
      "get": {
        "summary": "Lista todos os clientes",
        "tags": ["Clientes"],
        "responses": {
          "200": {
            "description": "Lista de clientes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Cliente"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria um cliente",
        "tags": ["Clientes"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteInput"
              },
              "example": {
                "nome": "Maria Silva",
                "email": "maria@email.com",
                "telefone": "(11) 99999-9999"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cliente"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/clientes/{id}": {
      "get": {
        "summary": "Busca um cliente por ID",
        "tags": ["Clientes"],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cliente"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Atualiza um cliente",
        "tags": ["Clientes"],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cliente atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cliente"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Remove um cliente",
        "tags": ["Clientes"],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "204": {
            "description": "Cliente removido"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/veiculos": {
      "get": {
        "summary": "Lista todos os veiculos",
        "tags": ["Veiculos"],
        "responses": {
          "200": {
            "description": "Lista de veiculos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Veiculo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria um veiculo",
        "tags": ["Veiculos"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VeiculoInput"
              },
              "example": {
                "placa": "ABC1D23",
                "marca": "Toyota",
                "modelo": "Corolla",
                "ano": 2020,
                "id_cliente": 1
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Veiculo criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Veiculo"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/veiculos/{id}": {
      "get": {
        "summary": "Busca um veiculo por ID",
        "tags": ["Veiculos"],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Veiculo encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Veiculo"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Atualiza um veiculo",
        "tags": ["Veiculos"],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VeiculoInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Veiculo atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Veiculo"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Remove um veiculo",
        "tags": ["Veiculos"],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "204": {
            "description": "Veiculo removido"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "ID do registro",
        "schema": {
          "type": "integer",
          "example": 1
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Dados invalidos ou obrigatorios ausentes",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiMessage"
            }
          }
        }
      },
      "NotFound": {
        "description": "Registro nao encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiMessage"
            }
          }
        }
      },
      "Conflict": {
        "description": "Operacao bloqueada por relacionamento ou regra de negocio",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiMessage"
            }
          }
        }
      }
    },
    "schemas": {
      "ApiMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Mensagem da API"
          }
        }
      },
      "Cliente": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              }
            }
          },
          {
            "$ref": "#/components/schemas/ClienteInput"
          }
        ]
      },
      "ClienteInput": {
        "type": "object",
        "required": ["nome", "email", "telefone"],
        "properties": {
          "nome": {
            "type": "string",
            "example": "Maria Silva"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "maria@email.com"
          },
          "telefone": {
            "type": "string",
            "example": "(11) 99999-9999"
          }
        }
      },
      "Veiculo": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              }
            }
          },
          {
            "$ref": "#/components/schemas/VeiculoInput"
          }
        ]
      },
      "VeiculoInput": {
        "type": "object",
        "required": ["placa", "marca", "modelo", "ano", "id_cliente"],
        "properties": {
          "placa": {
            "type": "string",
            "example": "ABC1D23"
          },
          "marca": {
            "type": "string",
            "example": "Toyota"
          },
          "modelo": {
            "type": "string",
            "example": "Corolla"
          },
          "ano": {
            "type": "integer",
            "example": 2020
          },
          "id_cliente": {
            "type": "integer",
            "example": 1
          }
        }
      },
...
    }
  }
}